rebind snippet completions and jumps
The previous supertab-like behaviour was causing issues. Sometimes it would snippet jump when I just meant to indent. I put everything on separate binds now. Jumping forwards and backwards with <C-h> and <C-l> seems to be a nonstandard keybind. It makes sense in my head since we scroll the snippet list with <C-j> and <C-k>.