nvim: disable semantic highlight for rust unresolved references
I might have to rethink this fix if I find more of these highlight
groups but for now it's the only one behaving like this.

Here's the problem:
An unresolved reference in rust is highlighted with red squiggles from
TWO separate sources: the lsp error, and the lsp semantic highlight.
My :NoiseToggle command is intended to toggle all distracting noise
about errors with the code itself. Before I ran into this, it had been
enough to just disable the lsp diagnostic highlights, but since this is
highlighted with red squiggles from the semantic highlight as well,
disabling the diagnostic highlight wasn't enough.

Why does it behave like this? Why does anyone want this as the default
behaviour? If you're not highlighting the lsp errors why the fuck would
you want semantic highlighting to still give you squiggles? It looks the
same to the 99% of users who don't just enable one of them.

Rust's LSP stuff behaves really different to everything else I've used
:P